move workflows.ts -> src/workflows.ts re: #13, add node-fetch to cancellable-fetch activity
