+++ cancellation/src/workflows/index.ts
@@ -1,8 +1,8 @@
 'use strict';
 
 import { createActivityHandle, isCancellation } from '@temporalio/workflow';
-import { Example } from './interfaces';
-import type * as activities from './activities';
+import { Example } from '../interfaces';
+import type * as activities from '../activities';
 
 const { activityToBeCancelled } = createActivityHandle<typeof activities>({
   startToCloseTimeout: '60s',
